third_party/blink/web_tests/fast/js/mozilla/strict/8.7.2-expected.txt

PASS 'use strict'; undeclared=1 threw exception of type ReferenceError.
PASS true === true
PASS true === true
PASS !!Function("'use strict'; undeclared_at_compiletime=1") is true
PASS !!Function("undeclared_at_compiletime=1") is true
PASS true === true
PASS with(obj()) { (function () { 'use strict'; x = 2; y = 2; })(); } threw exception of type TypeError.
PASS true === true
PASS with(obj()) { (function () { 'use strict'; x++; })(); } threw exception of type TypeError.
PASS true === true
PASS with(obj()) { (function () { 'use strict'; ++x; })(); } threw exception of type TypeError.
PASS true === true
PASS with(obj()) { (function () { 'use strict'; x--; })(); } threw exception of type TypeError.
PASS true === true
PASS with(obj()) { (function () { 'use strict'; --x; })(); } threw exception of type TypeError.
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
